pyproject.toml
[project]
name = "ai-interview-backed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.11"
dependencies = [
    "fastapi==0.110.0",
    "uvicorn==0.30.6",
    "sqlalchemy==1.4.46",
    "pydantic==2.10.6",
    "livekit==1.0.12",
    "python-jose==3.5.0",
    "passlib==1.7.4",
    "bcrypt==4.0.1",
    "alembic==1.8.1",
    "psycopg2-binary==2.9.10",
    "python-multipart==0.0.9",
]